rename the files now that the directive name has changed
